refactor(device): move set_temperature validation to AqualinkClimate base
IaquaClimate and ExoClimate had identical range-check logic. Base now
validates min_temp/max_temp/temperature_unit, dispatches to
_set_temperature. Concrete classes implement only the device command.
